Initial Commit of C# Socket Listener
This is intended to be a starting point for interacting with C# and go.
I've set up go to take in a request from '/jobs/crawler' and then hit a
different endpoint 'localhost:8001' that the c# will be listening on. No
data is being passed yet; I thought I would make sure I'm on the right
path first. Also, sorry about all the extra code.  Only thing you guys
should be concerned with is the Program.cs file